Correct rounding and add time part to date axes
1) Currently date rounding is always carried out to the nearest date. It
must be, however, carried out to the smallest time unit present, i.e. when
only yyyy is given as format string then rounding must be carried out to
the nearest year number. This commit implements correct rounding to the
smallest time unit.

2) Add time information to date axes, either in addition to date part
or on its own. The following rules apply:
- The placeholder for minutes is 'ii' as 'mm' is already used for months
  and the format string is not case sensitive.
- For both date and time parts, all bigger time units must be given starting
  from year or hour, respectively, down the to smallest unit needed (i.e.
  mm/dd or ii:ss are not valid, it must be yy/mm/dd or hh:ii:ss)
- When only the time part is used the date part is implicitely set to the
  current date.