small update on the documentation in .tex/.pdf (#295)
* use unbreakable space for calling figure labels

Spurred by section 5.3 where figure 12 is cross-referenced, now all
labels about figures are used with the fixed space (tilde sign).

* use unbreakable space if calling section labels

* update version information to current release

* update address about custom scripts

* adjust some TeX details in the documentation

For easier reading of the .tex, use of a blank line prior/after
an itemize/enumerate block.  In reference to IUPAC's Green Book,
use of roman subscripts because the subscripts do not describe a
physical property.

* add hint about git-lfs used in the project

The default installation of git yields an incompliet fork and local
check-out because the project uses git's large file storage.  This
requires an additional installation on top of git which now are
described.

* complete the revision